mobile-broadband-provider-info: Add bbappend to specify branch in SRC_URI
In the case of [1], do_fetch of mobile-broadband-provider-info of Poky fails.
This is because default branch name of mobile-broadband-provider-info's git
repository has changed to "main".

To fix the above issue, we add a bbappend file replacing the SRC_URI.

Referred commits in the Poky master branch:
- ddcf16d1f7 meta: Add explict branch to git SRC_URIs
- e4795393c4 mobile-broadband-provider-info: upgrade 20210805 -> 20220315

The meta-debian depends on Poky, but its warrior branch is not maintained.
There may potentially be other cases where unupdated Poky recipes cause build
errors similar to this one, and we expect to see more in the future. So we
introduce the "recipes-poky-compat" directory as a place to fix these issue.

[1]:
  Add the following to conf/local.conf:
  ```
  DISTRO = "deby"
  MACHINE = "qemuarm64"
  IMAGE_INSTALL_append = "connman"
  ```
  and run `bitbake core-image-minimal`, the following error:
  ```
  ERROR: mobile-broadband-provider-info-1_20190116-r0 do_fetch: Fetcher failure: Unable to find revision c7def60ba50d9cc30a90f69f89d7e82243501e86 in branch master even from upstream
  ERROR: mobile-broadband-provider-info-1_20190116-r0 do_fetch: Fetcher failure for URL: 'git://gitlab.gnome.org/GNOME/mobile-broadband-provider-info.git;protocol=https'. Unable to fetch URL from any source.
  ```

Signed-off-by: Takahiro Terada <takahiro.terada@miraclelinux.com>